Split headers and add slip.h and event.h
